adding role handling

+++ src/main/java/com/kosuri/stores/dao/SaleEntity.java
@@ -63,7 +63,7 @@ public class SaleEntity {
     @Column(name="Qty_Box",nullable = true, length = 45)
     private Double qtyBox;
     @Column(name="Qty",nullable = true, length = 45)
-    private Integer qty;
+    private Double qty;
     @Column(name="Sch_Qty",nullable = true, length = 45)
     private int schQty;
     @Column(name="Sch_Disc",nullable = true, length = 45)
@@ -241,10 +241,10 @@ public class SaleEntity {
     public void setQtyBox(Double qtyBox) {
         this.qtyBox = qtyBox;
     }
-    public Integer getQty() {
+    public Double getQty() {
         return qty;
     }
-    public void setQty(Integer qty) {
+    public void setQty(Double qty) {
         this.qty = qty;
     }
     public int getSchQty() {
